chore: consolidated constants & removed commented code
